--- benches/benchmarks/mod.rs
@@ -1,10 +1,0 @@
-#![allow(dead_code)]
-
-#[cfg(feature = "gaiku_baker_heightmap")]
-pub mod heightmap;
-
-#[cfg(feature = "gaiku_baker_marching_cubes")]
-pub mod marching_cubes;
-
-#[cfg(feature = "gaiku_baker_voxel")]
-pub mod voxel;
